fix some new build warnings

+++ nav2_behavior_tree/test/plugins/condition/test_are_error_codes_present.cpp
@@ -27,8 +27,8 @@ public:
   using ActionGoal = Action::Goal;
   void SetUp()
   {
-    int error_code = ActionGoal::NONE;
-    std::set<unsigned short> error_codes_to_check = {ActionGoal::UNKNOWN}; //NOLINT
+    unsigned int error_code = ActionGoal::NONE;
+    std::set<unsigned int> error_codes_to_check = {ActionGoal::UNKNOWN}; //NOLINT
     config_->blackboard->set("error_code", error_code);
     config_->blackboard->set("error_codes_to_check", error_codes_to_check);
